feat(front) : WIP table / table-legacy
