+++ packages/scss/src/components/tableFixed/mods.scss
@@ -1,5 +1,24 @@
-@mixin layoutFixed($i) {
-	min-width: $i * 1rem;
-	max-width: $i * 1rem;
-	width: $i * 1rem;
+@use '@lucca-front/scss/src/commons/config';
+@use '@lucca-front/scss/src/commons/utils/media';
+
+@mixin layoutFixed {
+    [class*='row-cell'] {
+        --cell-width: calc(var(--table-layout-fixed-width) * 1rem);
+    }
+    &.mod-layoutFixed [class*='row-cell'] {
+        @include cellFixedWidth;
+    }
+    @each $breakpoint, $value in config.$breakpoints {
+        @include media.min($breakpoint) {
+            &.mod-layoutFixed\@mediaMin#{$breakpoint} [class*='row-cell'] {
+                @include cellFixedWidth;
+            }
+        }
+    }
+}
+
+@mixin cellFixedWidth {
+    min-width: var(--cell-width, auto);
+    max-width: var(--cell-width, auto);
+    width: var(--cell-width, auto);
 }
