+++ packages/scss/src/components/tableFixedLegacy/exports.scss
