--- packages/scss/src/components/tableFixedV2/index.scss
@@ -1,5 +1,0 @@
-@use 'exports' as *;
-
-[class*='mod-layoutFixedV2'] {
-	@include layoutFixed;
-}
