--- packages/scss/src/components/tableFixedV2/mods.scss
@@ -1,24 +1,0 @@
-@use '@lucca-front/scss/src/commons/config';
-@use '@lucca-front/scss/src/commons/utils/media';
-
-@mixin layoutFixed {
-    [class*='row-cell'] {
-        --cell-width: calc(var(--table-layout-fixed-width) * 1rem);
-    }
-    &.mod-layoutFixedV2 [class*='row-cell'] {
-        @include cellFixedWidth;
-    }
-    @each $breakpoint, $value in config.$breakpoints {
-        @include media.min($breakpoint) {
-            &.mod-layoutFixedV2\@mediaMin#{$breakpoint} [class*='row-cell'] {
-                @include cellFixedWidth;
-            }
-        }
-    }
-}
-
-@mixin cellFixedWidth {
-    min-width: var(--cell-width, auto);
-    max-width: var(--cell-width, auto);
-    width: var(--cell-width, auto);
-}
