+++ packages/scss/src/components/tableSticked/vars.scss
@@ -1,8 +1,4 @@
 @mixin vars {
-	// --components-table-sticky-column-max-offset: 50;
-	// --components-table-fixed-column-min-col-width: 2;
-	// --components-table-fixed-column-max-col-width: 20;
-	// --components-table-fixed-column-sticky-shadow-color: #2a3551;
-
-	--components-table-fixed-column-sticky-shadow-width: 0.625rem;
+	--components-table-sticked-column-sticky-shadow-width: 0.625rem;
+	--components-table-sticked-column-sticky-shadow-color: rgba(42, 53, 81, 0.25);
 }
