+++ packages/scss/src/components/tableStickedLegacy/exports.scss
