+++ packages/scss/src/components/tableStickedLegacy/index.scss
@@ -1,0 +1,58 @@
+@use '@lucca-front/scss/src/commons/config';
+@use '@lucca-front/scss/src/commons/utils/media';
+@use 'exports' as *;
+
+.table {
+	@include vars;
+
+	&.mod-stickyColumn {
+		@include stickyColumn;
+	}
+
+	@each $breakpoint, $value in config.$breakpoints {
+		@include media.min($breakpoint) {
+			&.mod-stickyColumn\@mediaMin#{$breakpoint} {
+				@include stickyColumn;
+			}
+		}
+	}
+
+	&[class*='mod-stickyColumn'] {
+		.table-head-row-cell,
+		.table-body-row-cell,
+		.table-foot-row-cell {
+			@for $i from 0 through 50 {
+				&[class*='mod-stickyColumn'][class*='Offset#{$i}'] {
+					@include stickyOffset($i);
+				}
+			}
+		}
+	}
+
+	&[class*='mod-stickyHeader'] {
+		@include stickyHeaderCommon;
+
+		@each $breakpoint, $value in config.$breakpoints {
+			@include media.max($breakpoint) {
+				&.mod-stickyColumn\@mediaMin#{$breakpoint} {
+					@include stickyColumnBreakpoint;
+				}
+			}
+		}
+	}
+
+	&.mod-stickyHeader {
+		@include stickyHeader;
+	}
+
+	&.mod-stickyHeader-withBanner {
+		@include stickyHeaderBanner;
+	}
+}
+
+.table-body-row,
+.table-foot-row {
+	&.mod-stickyHeader-shadow {
+		@include stickyHeaderShadow;
+	}
+}
