+++ packages/scss/src/components/tableStickedLegacy/vars.scss
@@ -1,0 +1,8 @@
+@mixin vars {
+	// --components-table-sticky-column-max-offset: 50;
+	// --components-table-fixed-column-min-col-width: 2;
+	// --components-table-fixed-column-max-col-width: 20;
+	// --components-table-fixed-column-sticky-shadow-color: #2a3551;
+
+	--components-table-fixed-column-sticky-shadow-width: 0.625rem;
+}
