--- packages/scss/src/components/tableStickedV2/index.scss
@@ -1,41 +1,0 @@
-@use '@lucca-front/scss/src/commons/config';
-@use '@lucca-front/scss/src/commons/utils/media';
-@use 'exports' as *;
-
-.table {
-	@include vars;
-
-	//For both mod-stickyColumnV2 + responsive variant
-	&[class*='mod-stickyColumnV2'] {
-		@include stickyColumn;
-	}
-
-	.mod-stickyColumnV2 {
-		@include stickyColumnOffset;
-	}
-
-	@each $breakpoint, $value in config.$breakpoints {
-		@include media.min($breakpoint) {
-			&.mod-stickyColumnV2\@mediaMin#{$breakpoint} {
-				@include stickyColumnOffset;
-			}
-		}
-	}
-
-	//For both stickyHeaderV2 + -withBanner variant
-	&[class*='mod-stickyHeaderV2'] {
-		@include stickyHeaderCommon;
-
-		.mod-stickyHeader-shadow {
-			@include stickyHeaderShadow;
-		}
-	}
-
-	&.mod-stickyHeaderV2 {
-		@include stickyHeader;
-	}
-
-	&.mod-stickyHeaderV2-withBanner {
-		@include stickyHeaderBanner;
-	}
-}
