--- packages/scss/src/components/tableStickedV2/mods.scss
@@ -1,187 +1,0 @@
-@use 'sass:color';
-@use '@lucca-front/scss/src/commons/utils/media';
-@use '@lucca-front/scss/src/commons/config';
-@use '@lucca-front/icons/src/commons/utils/icon';
-@use '@lucca-front/scss/src/commons/utils/reset';
-
-@mixin stickyColumn {
-	width: auto;
-	min-width: 100%;
-	background-color: var(--colors-white-color);
-
-	//All stickies columns
-	[class*='mod-stickyColumn-'] {
-		background-color: var(--colors-white-color);
-		z-index: 3;
-	}
-
-	//Left sticked columns
-	[class*='mod-stickyColumn-left'] {
-		left: var(--components-table-sticked-column-sticky-offset);
-
-		//left sticked columns shadow
-		&.mod-stickyColumn-shadow,
-		.stickyColumn-shadow-wrapper::before {
-			left: calc(
-				var(--components-table-sticked-column-sticky-offset) -
-					var(--components-table-sticked-column-sticky-shadow-width)
-			);
-		}
-		.stickyColumn-shadow-wrapper::after {
-			left: var(--components-table-sticked-column-sticky-offset);
-			background-image: linear-gradient(
-				to right,
-				var(--components-table-sticked-column-sticky-shadow-color),
-				transparent
-			);
-		}
-	}
-
-	//Right sticked columns
-	[class*='mod-stickyColumn-right'] {
-		right: var(--components-table-sticked-column-sticky-offset);
-
-		//right sticked columns shadow
-		.stickyColumn-shadow-wrapper {
-			justify-items: end;
-			width: calc(var(--components-table-sticked-column-sticky-shadow-width) * 3);
-			right: calc(var(--components-table-sticked-column-sticky-shadow-width) * -1);
-		}
-		&.mod-stickyColumn-shadow,
-		.stickyColumn-shadow-wrapper::before {
-			right: calc(
-				var(--components-table-sticked-column-sticky-offset) -
-					var(--components-table-sticked-column-sticky-shadow-width)
-			);
-		}
-		.stickyColumn-shadow-wrapper::after {
-			right: var(--components-table-sticked-column-sticky-offset);
-			background-image: linear-gradient(
-				to left,
-				var(--components-table-sticked-column-sticky-shadow-color),
-				transparent
-			);
-		}
-	}
-
-	//Sticky columns drop shadow
-	.mod-stickyColumn-shadow {
-		display: none;
-		width: 0;
-		min-width: 0;
-		padding: 0;
-	}
-
-	.stickyColumn-shadow-wrapper {
-		position: absolute;
-		top: 0;
-		bottom: 0;
-		display: grid;
-		grid-template-columns: calc(var(--components-table-sticked-column-sticky-shadow-width) * 2);
-		grid-template-areas: 'cell';
-		width: calc(var(--components-table-sticked-column-sticky-shadow-width) * 2);
-
-		&::before,
-		&::after {
-			content: '';
-			position: sticky;
-			grid-area: cell;
-		}
-		&::after {
-			width: var(--components-table-sticked-column-sticky-shadow-width);
-		}
-		&::before {
-			z-index: 1;
-			width: var(--components-table-sticked-column-sticky-shadow-width);
-			background-color: var(--colors-white-color);
-		}
-	}
-}
-
-@mixin stickyColumnOffset {
-	[class*='mod-stickyColumn-'] {
-		position: sticky;
-		--components-table-sticked-column-sticky-offset: calc(var(--table-sticky-column-offset) * 1rem);
-	}
-	.mod-stickyColumn-shadow {
-		display: table-cell;
-	}
-}
-
-/*****
- Sticky header
-****/
-
-@mixin stickyHeader {
-	.table-head-row-cell {
-		top: calc(var(--components-table-sticked-column-sticky-shadow-width) * -1);
-	}
-}
-
-@mixin stickyHeaderBanner {
-	.table-head-row-cell {
-		top: var(--commons-banner-height);
-	}
-	.mod-stickyHeader-shadow .table-body-row-cell {
-		top: calc(
-			var(--table-sticky-header-shadow-offset) + var(--commons-banner-height) +
-				var(--components-table-sticked-column-sticky-shadow-width)
-		);
-	}
-}
-
-@mixin stickyHeaderCommon {
-	margin-top: var(--components-table-sticked-column-sticky-shadow-width);
-	.table-head-row-cell {
-		background-color: var(--colors-white-color);
-		position: sticky;
-		z-index: 5;
-		&[class*='mod-stickyColumn'] {
-			z-index: 7;
-		}
-	}
-}
-
-/*****
- Sticky header drop shadow
-****/
-
-@mixin stickyHeaderShadow {
-	[class*='row-cell'] {
-		position: sticky;
-		top: var(--table-sticky-header-shadow-offset);
-		z-index: 4;
-		height: 0;
-		padding: 0;
-		border: 0;
-		background: transparent;
-	}
-
-	.stickyHeader-shadow-wrapper {
-		position: absolute;
-		top: calc(var(--components-table-sticked-column-sticky-shadow-width) * -1);
-		width: 100%;
-		height: 0;
-		border: 0;
-
-		&::after {
-			content: '';
-			display: block;
-			height: var(--components-table-sticked-column-sticky-shadow-width);
-			width: 100%;
-			background-image: linear-gradient(
-				to bottom,
-				var(--components-table-sticked-column-sticky-shadow-color),
-				transparent
-			);
-			opacity: 0.5;
-		}
-	}
-
-	+ .table-body-row,
-	+ .table-foot-row {
-		[class*='row-cell'] {
-			border-top: 0;
-		}
-	}
-}
