--- packages/scss/src/components/tableStickedV2/vars.scss
@@ -1,4 +1,0 @@
-@mixin vars {
-	--components-table-sticked-column-sticky-shadow-width: 0.625rem;
-	--components-table-sticked-column-sticky-shadow-color: rgba(42, 53, 81, 0.25);
-}
