+++ stories/documentation/listings/table-legacy/table-legacy-sticky-columns-breakpoints.stories.ts
@@ -1,30 +1,28 @@
-import { Meta, Story } from '@storybook/angular';
+import { Meta, StoryFn } from '@storybook/angular';
 
-interface TableStickyColumnsAndHeaderWithBreakpointsStory {
-}
+interface TableLegacyStickyColumnsAndHeaderWithBreakpointsStory { }
 
 export default {
-	title: 'Documentation/Listings/Table/Sticky Columns And Header With Breakpoints v2',
-	argTypes: {
-	},
+	title: 'Documentation/Listings/Table/Legacy/Sticky Columns And Header With Breakpoints',
+	argTypes: {},
 } as Meta;
 
-function getTemplate(args: TableStickyColumnsAndHeaderWithBreakpointsStory): string {
+function getTemplate(args: TableLegacyStickyColumnsAndHeaderWithBreakpointsStory): string {
 	return `
 	<div class="demo-wrapper">
-		<table class="table mod-layoutFixedV2@mediaMinS mod-stickyColumnV2@mediaMinS mod-stickyHeaderV2" style="--table-sticky-header-shadow-offset: 56px">
+		<table class="table mod-layoutFixed@mediaMinXS mod-stickyColumn@mediaMinXS mod-stickyHeader">
 			<thead class="table-head">
 				<tr class="table-head-row">
-					<th class="table-head-row-cell mod-stickyColumn-left" style="--table-layout-fixed-width: 8; --table-sticky-column-offset: 0">
+					<th class=" table-head-row-cell mod-layoutFixed-8@mediaMinXS mod-stickyColumn-leftOffset0 " >
 						Head cell
 					</th>
-					<th class="table-head-row-cell mod-stickyColumn-left" style="--table-layout-fixed-width: 7; --table-sticky-column-offset: 8">
+					<th class=" table-head-row-cell mod-layoutFixed-7@mediaMinXS mod-stickyColumn-leftOffset8 " >
 						Head cell
 					</th>
-					<th class="table-head-row-cell mod-stickyColumn-left" style="--table-layout-fixed-width: 5; --table-sticky-column-offset: 15">
+					<th class=" table-head-row-cell mod-layoutFixed-5@mediaMinXS mod-stickyColumn-leftOffset15 " >
 						Head cell
 					</th>
-					<th class="table-head-row-cell mod-stickyColumn-left mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 20">
+					<th class=" table-head-row-cell mod-stickyColumn-leftOffset20 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</th>
 					<th class="table-head-row-cell">Head cell</th>
@@ -37,31 +35,31 @@ function getTemplate(args: TableStickyColumnsAndHeaderWithBreakpointsStory): str
 					<th class="table-head-row-cell">Head cell</th>
 					<th class="table-head-row-cell">Head cell</th>
 					<th class="table-head-row-cell">Head cell</th>
-					<th class="table-head-row-cell mod-stickyColumn-right mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 5">
+					<th class=" table-head-row-cell mod-stickyColumn-rightOffset5 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</th>
-					<th class="table-head-row-cell mod-stickyColumn-right" style="--table-layout-fixed-width: 5; --table-sticky-column-offset: 0">
+					<th class=" table-head-row-cell mod-layoutFixed-5@mediaMinXS mod-stickyColumn-rightOffset0 " >
 						Head cell
 					</th>
 				</tr>
 			</thead>
 			<tbody class="table-body">
 				<tr class="table-body-row mod-stickyHeader-shadow">
-					<td class="table-body-row-cell" colspan="16" role="presentation">
+					<td class="table-body-row-cell" style="--sticky-header-shadow-offset-top: 56px" colspan="16" role="presentation" >
 						<div class="stickyHeader-shadow-wrapper"></div>
 					</td>
 				</tr>
 				<tr class="table-body-row">
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 0">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset0">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 8">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset8">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 15">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset15">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 20">
+					<td class=" table-body-row-cell mod-stickyColumn-leftOffset20 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</td>
 					<td class="table-body-row-cell">Body cell</td>
@@ -74,24 +72,24 @@ function getTemplate(args: TableStickyColumnsAndHeaderWithBreakpointsStory): str
 					<td class="table-body-row-cell">Body cell</td>
 					<td class="table-body-row-cell">Body cell</td>
 					<td class="table-body-row-cell">Body cell</td>
-					<td class="table-body-row-cell mod-stickyColumn-right mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 5">
+					<td class=" table-body-row-cell mod-stickyColumn-rightOffset5 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-right" style="--table-sticky-column-offset: 0">
+					<td class="table-body-row-cell mod-stickyColumn-rightOffset0">
 						Body cell
 					</td>
 				</tr>
 				<tr class="table-body-row">
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 0">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset0">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 8">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset8">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 15">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset15">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 20">
+					<td class=" table-body-row-cell mod-stickyColumn-leftOffset20 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</td>
 					<td class="table-body-row-cell">Body cell</td>
@@ -104,24 +102,24 @@ function getTemplate(args: TableStickyColumnsAndHeaderWithBreakpointsStory): str
 					<td class="table-body-row-cell">Body cell</td>
 					<td class="table-body-row-cell">Body cell</td>
 					<td class="table-body-row-cell">Body cell</td>
-					<td class="table-body-row-cell mod-stickyColumn-right mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 5">
+					<td class=" table-body-row-cell mod-stickyColumn-rightOffset5 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-right" style="--table-sticky-column-offset: 0">
+					<td class="table-body-row-cell mod-stickyColumn-rightOffset0">
 						Body cell
 					</td>
 				</tr>
 				<tr class="table-body-row">
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 0">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset0">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 8">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset8">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left" style="--table-sticky-column-offset: 15">
+					<td class="table-body-row-cell mod-stickyColumn-leftOffset15">
 						Body cell
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-left mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 20">
+					<td class=" table-body-row-cell mod-stickyColumn-leftOffset20 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</td>
 					<td class="table-body-row-cell">Body cell</td>
@@ -134,25 +132,23 @@ function getTemplate(args: TableStickyColumnsAndHeaderWithBreakpointsStory): str
 					<td class="table-body-row-cell">Body cell</td>
 					<td class="table-body-row-cell">Body cell</td>
 					<td class="table-body-row-cell">Body cell</td>
-					<td class="table-body-row-cell mod-stickyColumn-right mod-stickyColumn-shadow" role="presentation" style="--table-sticky-column-offset: 5">
+					<td class=" table-body-row-cell mod-stickyColumn-rightOffset5 mod-stickyColumn-shadow " role="presentation" >
 						<div class="stickyColumn-shadow-wrapper"></div>
 					</td>
-					<td class="table-body-row-cell mod-stickyColumn-right" style="--table-sticky-column-offset: 0">
+					<td class="table-body-row-cell mod-stickyColumn-rightOffset0">
 						Body cell
 					</td>
 				</tr>
 			</tbody>
 		</table>
 	</div>
-	`
+	`;
 }
 
-const Template: Story<TableStickyColumnsAndHeaderWithBreakpointsStory> = (args: TableStickyColumnsAndHeaderWithBreakpointsStory) => ({
+const Template: StoryFn<TableLegacyStickyColumnsAndHeaderWithBreakpointsStory> = (args: TableLegacyStickyColumnsAndHeaderWithBreakpointsStory) => ({
 	props: args,
 	template: getTemplate(args),
-	styles: [
-		`.demo-wrapper {overflow: auto; height: 10rem;}`
-	]
+	styles: [`.demo-wrapper {overflow: auto; height: 10rem;}`],
 });
 
 export const StickyColumnsAndHeaderWithBreakpoints = Template.bind({});
