+++ stories/qa/table/table.stories.html
@@ -701,15 +701,15 @@
 	<table class="table mod-layoutFixed">
 		<thead class="table-head">
 			<tr class="table-head-row">
-				<th class="table-head-row-cell mod-layoutFixed-8">Head cell</th>
-				<th class="table-head-row-cell mod-layoutFixed-12">Head cell</th>
+				<th class="table-head-row-cell" style="--table-layout-fixed-width: 8;">Fixed 8rem head cell</th>
+				<th class="table-head-row-cell" style="--table-layout-fixed-width: 12;">Fixed 12rem head cell</th>
 				<th class="table-head-row-cell">Head cell</th>
 			</tr>
 		</thead>
 		<tbody class="table-body">
 			<tr class="table-body-row">
-				<td class="table-body-row-cell">Cell with mod-layoutFixed-8</td>
-				<td class="table-body-row-cell">Cell with mod-layoutFixed-12</td>
+				<td class="table-body-row-cell">Body cell</td>
+				<td class="table-body-row-cell">Body cell</td>
 				<td class="table-body-row-cell">Body cell</td>
 			</tr>
 			<tr class="table-body-row">
@@ -875,6 +875,7 @@
 		previous cells in the specified axis (from left or from right)</em>
 </section>
 
+<!-- TODO a modif-->
 <!-- Sticky columns with shadow-->
 <section class="contentSection">
 	<h2>Sticky columns with shadow on overflow</h2>
